DI Antipatterns
