+++ aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample/Startup.cs
@@ -20,7 +20,7 @@ namespace DependencyInjectionSample
             services.AddTransient<IOperationTransient, Operation>();
             services.AddScoped<IOperationScoped, Operation>();
             services.AddSingleton<IOperationSingleton, Operation>();
-            services.AddSingleton<IOperationSingletonInstance>(new Operation(Guid.Empty));
+            services.AddSingleton<IOperationSingletonInstance>(new Operation("00:00.0"));
 
             // OperationService depends on each of the other Operation types.
             services.AddTransient<OperationService, OperationService>();
