--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/DependencyInjectionSample.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-  
-</Project>
