--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Interfaces/IMyDependency.cs
@@ -1,11 +1,0 @@
-using System.Threading.Tasks;
-
-namespace DependencyInjectionSample.Interfaces
-{
-    #region snippet1
-    public interface IMyDependency
-    {
-        Task WriteMessage(string message);
-    }
-    #endregion
-}
