--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Interfaces/IOperation.cs
@@ -1,27 +1,0 @@
-using System;
-
-namespace DependencyInjectionSample.Interfaces
-{
-    #region snippet1
-    public interface IOperation
-    {
-        string OperationId { get; }
-    }
-
-    public interface IOperationTransient : IOperation
-    {
-    }
-
-    public interface IOperationScoped : IOperation
-    {
-    }
-
-    public interface IOperationSingleton : IOperation
-    {
-    }
-
-    public interface IOperationSingletonInstance : IOperation
-    {
-    }
-    #endregion
-}
