--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Models/Operation.cs
@@ -1,35 +1,0 @@
-using System;
-using DependencyInjectionSample.Interfaces;
-
-namespace DependencyInjectionSample.Models
-{
-    #region snippet1
-    public class Operation : IOperationTransient, 
-        IOperationScoped, 
-        IOperationSingleton, 
-        IOperationSingletonInstance
-    {
-        public Operation() : this(DateTime.Now.ToString("hh:mm:ss.fff"))
-        {
-            DoWork();
-        }
-
-        public Operation(string id)
-        {
-            OperationId = id;
-        }
-
-        private void DoWork()
-        {
-            // Do work so different times are recorded.
-            int sum = 0;
-            for (int i = 0; i < 1000000; i++)
-            {
-                sum += i;
-            }
-        }
-
-        public string OperationId { get; private set; }
-    }
-    #endregion
-}
