--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Pages/Index.cshtml
@@ -1,50 +1,0 @@
-﻿@page
-@model IndexModel
-@{
-    ViewData["Title"] = "Dependency Injection Sample";
-}
-
-<h1>@ViewData["Title"]</h1>
-
-<div class="row">
-    <div class="panel panel-default">
-        <div class="panel-heading">
-            <h2 class="panel-title">MyDependency</h2>
-        </div>
-        <div class="panel-body">
-            <p>The <code>IndexModel</code> has called the <code>IMyDependency</code> implemention (<code>MyDependency</code>). The service's <code>WriteMessage</code> method was called with a message string that indicates the page called the method on the service. The <code>WriteMessage</code> method created a log entry with the message string.</p>
-        </div>
-    </div>
-</div>
-
-<div class="row">
-    <div class="panel panel-default">
-        <div class="panel-heading">
-            <h2 class="panel-title">Operations</h2>
-        </div>
-        <div class="panel-body">
-            <h3>Page Model Operations</h3>
-            <dl>
-                <dt>Transient</dt>
-                <dd>@Model.TransientOperation.OperationId</dd>
-                <dt>Scoped</dt>
-                <dd>@Model.ScopedOperation.OperationId</dd>
-                <dt>Singleton</dt>
-                <dd>@Model.SingletonOperation.OperationId</dd>
-                <dt>Instance</dt>
-                <dd>@Model.SingletonInstanceOperation.OperationId</dd>
-            </dl>
-            <h3>OperationService Operations</h3>
-            <dl>
-                <dt>Transient</dt>
-                <dd>@Model.OperationService.TransientOperation.OperationId</dd>
-                <dt>Scoped</dt>
-                <dd>@Model.OperationService.ScopedOperation.OperationId</dd>
-                <dt>Singleton</dt>
-                <dd>@Model.OperationService.SingletonOperation.OperationId</dd>
-                <dt>Instance</dt>
-                <dd>@Model.OperationService.SingletonInstanceOperation.OperationId</dd>
-            </dl>
-        </div>
-    </div>
-</div>
