--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Pages/Index2.cshtml
@@ -1,21 +1,0 @@
-﻿@page
-@model Index2Model
-@{
-    ViewData["Title"] = "Dependency Injection Sample";
-}
-
-<h1>@ViewData["Title"]</h1>
-
-<div class="row">
-    <div class="panel panel-default">
-        <div class="panel-heading">
-            <h2 class="panel-title">MyDependency</h2>
-        </div>
-        <div class="panel-body">
-            <p>The <code>IndexModel</code> has called the <code>IMyDependency</code> implemention
-            (<code>MyDependency</code>). The service's <code>WriteMessage</code> method was 
-            called with a message string that indicates the page called the method on the service.
-            The <code>WriteMessage</code> method created a log entry with the message string.</p>
-        </div>
-    </div>
-</div>
