--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Pages/Index2.cshtml.cs
@@ -1,25 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using DependencyInjectionSample.Interfaces;
-using DependencyInjectionSample.Services;
-
-namespace DependencyInjectionSample.Pages
-{
-    #region snippet1
-    public class Index2Model : PageModel
-    {
-        private readonly IMyDependency _myDependency;
-
-        public Index2Model(IMyDependency myDependency)
-        {
-            _myDependency = myDependency;            
-        }
-
-        public async Task OnGetAsync()
-        {
-            await _myDependency.WriteMessage(
-                "IndexModel.OnGetAsync created this message.");
-        }
-    }
-    #endregion
-}
