--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Pages/_ViewImports.cshtml
@@ -1,4 +1,0 @@
-@using DependencyInjectionSample
-@using DependencyInjectionSample.Models
-@namespace DependencyInjectionSample.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
