--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Pages/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
