--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/README.md
@@ -1,3 +1,0 @@
-# ASP.NET Core Dependency Injection
-
-This sample illustrates the use of dependency injection with ASP.NET Core. This sample demonstrates the scenario described in the [Dependency injection in ASP.NET Core](https://docs.microsoft.com/aspnet/core/fundamentals/dependency-injection) topic.
