--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Services/MyDependency.cs
@@ -1,27 +1,0 @@
-using System.Threading.Tasks;
-using DependencyInjectionSample.Interfaces;
-using Microsoft.Extensions.Logging;
-
-namespace DependencyInjectionSample.Services
-{
-    #region snippet1
-    public class MyDependency : IMyDependency
-    {
-        private readonly ILogger<MyDependency> _logger;
-
-        public MyDependency(ILogger<MyDependency> logger)
-        {
-            _logger = logger;
-        }
-
-        public Task WriteMessage(string message)
-        {
-            _logger.LogInformation(
-                "MyDependency.WriteMessage called. Message: {MESSAGE}", 
-                message);
-
-            return Task.FromResult(0);
-        }
-    }
-    #endregion
-}
