--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Services/OperationService.cs
@@ -1,26 +1,0 @@
-using DependencyInjectionSample.Interfaces;
-
-namespace DependencyInjectionSample.Services
-{
-    #region snippet1
-    public class OperationService
-    {
-        public OperationService(
-            IOperationTransient transientOperation,
-            IOperationScoped scopedOperation,
-            IOperationSingleton singletonOperation,
-            IOperationSingletonInstance instanceOperation)
-        {
-            TransientOperation = transientOperation;
-            ScopedOperation = scopedOperation;
-            SingletonOperation = singletonOperation;
-            SingletonInstanceOperation = instanceOperation;
-        }
-
-        public IOperationTransient TransientOperation { get; }
-        public IOperationScoped ScopedOperation { get; }
-        public IOperationSingleton SingletonOperation { get; }
-        public IOperationSingletonInstance SingletonInstanceOperation { get; }
-    }
-    #endregion
-}
