--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/Startup.cs
@@ -1,51 +1,0 @@
-using System;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using DependencyInjectionSample.Interfaces;
-using DependencyInjectionSample.Models;
-using DependencyInjectionSample.Services;
-
-namespace DependencyInjectionSample
-{
-    public class Startup
-    {
-        #region snippet1
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddRazorPages();
-
-            services.AddScoped<IMyDependency, MyDependency>();
-            services.AddTransient<IOperationTransient, Operation>();
-            services.AddScoped<IOperationScoped, Operation>();
-            services.AddSingleton<IOperationSingleton, Operation>();
-            services.AddSingleton<IOperationSingletonInstance>(new Operation("OOOO"));
-
-            // OperationService depends on each of the other Operation types.
-            services.AddTransient<OperationService, OperationService>();
-        }
-        #endregion
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Error");
-            }
-
-            app.UseStaticFiles();
-
-            app.UseRouting();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapRazorPages();
-            });
-        }
-    }
-}
