--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/StartupMyDependency.cs
@@ -1,42 +1,0 @@
-using DependencyInjectionSample.Interfaces;
-using DependencyInjectionSample.Services;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace DependencyInjectionSample
-{
-    public class StartupMyDependency
-    {
-        #region snippet1
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddRazorPages();
-
-            services.AddScoped<IMyDependency, MyDependency>();
-        }
-        #endregion
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Error");
-            }
-
-            app.UseStaticFiles();
-
-            app.UseRouting();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapRazorPages();
-            });
-        }
-    }
-}
