--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/appsettings.Development.json
@@ -1,9 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Debug",
-      "System": "Information",
-      "Microsoft": "Information"
-    }
-  }
-}
