--- aspnetcore/fundamentals/dependency-injection/samples/3.x/DependencyInjectionSample2/appsettings.json
@@ -1,10 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*"
-}
